requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.16.0
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
whitenoise==5.2.0
